refactoring

+++ resources/views/task_statuses/index.blade.php
@@ -24,8 +24,12 @@
                         <td class="col-md-2">
                             <div class="container">
                                 <div class="row">
-                                    <a href="{{ route('task_statuses.edit', $status) }}" class="btn btn-primary col-6">{{ __('Edit') }}</a>
-                                    <a href="{{ route('task_statuses.destroy', $status) }}" data-confirm="Are you sure?" data-method="delete" rel="nofollow" class="btn btn-primary col-6">{{ __('Delete') }}</a>
+                                    @can('update', $status)
+                                        <a href="{{ route('task_statuses.edit', $status) }}" class="btn btn-primary col-6">{{ __('Edit') }}</a>
+                                    @endcan
+                                    @can('delete', $status)
+                                        <a href="{{ route('task_statuses.destroy', $status) }}" data-confirm="Are you sure?" data-method="delete" rel="nofollow" class="btn btn-primary col-6">{{ __('Delete') }}</a>
+                                    @endcan
                                 </div>
                             </div>
                         </td>
